fix: Zod eta validation — remove .int() from ProgressUpdateSchema
eta is computed as Date.now() + float arithmetic, yielding a non-integer
ms timestamp (e.g. 1713196864030.612). The .int() constraint rejected every
progress update containing an ETA, returning 400 and silently dropping all
progress events → progress bar frozen on Main.

+++ packages/shared/src/schemas.ts
@@ -51,7 +51,7 @@ export const ProgressUpdateSchema = z.object({
   workerId: z.string().min(1),
   progress: z.number().int().min(0).max(100),
   fps:      z.number().min(0).optional(),
-  eta:      z.number().int().optional(),
+  eta:      z.number().optional(),          // ms timestamp — may be a float
   phase:    z.enum(['receiving', 'transcoding', 'sending', 'swapping']),
 });
